% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBern1dep.R
\name{cBern1dep}
\alias{cBern1dep}
\title{Generate binary data with 1-dependent correlated structure}
\usage{
cBern1dep(n, p, rho)
}
\arguments{
\item{n}{number of observations}

\item{p}{the vector of marginal probabilities with dimension m}

\item{rho}{either a non-negative value indecating the shared correlation coefficient
or and m-1 vector indicating the correlation coefficients between adjacent
variables.}
}
\value{
an n*p matrix of binary data
}
\description{
Equivalent to cBern(n, p, rho, type="1-dependent")
}
\examples{
X <- cBern1dep(5, c(0.4,0.5,0.6), c(0.2,0.3))
}
