% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBernDCP.R
\name{cBernDCP}
\alias{cBernDCP}
\title{Generate binary data with decaying-product correlated structure}
\usage{
cBernDCP(n, p, rho)
}
\arguments{
\item{n}{number of observations}

\item{p}{the vector of marginal probabilities with dimension m}

\item{rho}{either a non-negative value indecating the shared correlation coefficient
or and m-1 vector indicating the correlation coefficients between adjacent
variables.}
}
\value{
an n*p matrix of binary data
}
\description{
Equivalent to cBern(n, p, rho, type="DCP")
}
\examples{
X <- cBernDCP(10, rep(0.5,3), c(0.2,0.2))
}
