% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoMaxMdep.R
\name{rhoMax1dep}
\alias{rhoMax1dep}
\title{To calculate the maximal allowed correlations
max for using cBern1dep to generate binary data
with 1-dependent structure}
\usage{
rhoMax1dep(p)
}
\arguments{
\item{p}{the vector of marginal probabilities with dimension m}
}
\value{
an (m-1)-dimensional vector rho, which is the maximum
the correlation between the adjacent variables
}
\description{
To calculate the maximal allowed correlations
max for using cBern1dep to generate binary data
with 1-dependent structure
}
