% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoMaxDCP.R
\name{rhoMaxDCP}
\alias{rhoMaxDCP}
\title{For calculating the maximal allowed correlations
max for binary data with decaying-product structure.}
\usage{
rhoMaxDCP(p)
}
\arguments{
\item{p}{marginal probabilities}
}
\value{
an (m-1)-dimensional vector rho, which is the maximum
the correlation between the adjacent variables
}
\description{
For calculating the maximal allowed correlations
max for binary data with decaying-product structure.
}
