% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoMaxEx.R
\name{rhoMaxEx}
\alias{rhoMaxEx}
\title{For calculating the maximal allowed correlation
coefficient for binary data with exchangeable structure.}
\usage{
rhoMaxEx(p)
}
\arguments{
\item{p}{the vector of marginal probabilities with dimension m}
}
\value{
the maximal allowed correlation coefficient
}
\description{
For calculating the maximal allowed correlation
coefficient for binary data with exchangeable structure.
}
